Implemented logs for Cover Image
